+++ R/aaa.R
@@ -1,0 +1,2 @@
+
+`_altair_chunk_ids` <- new.env(parent = emptyenv())