+++ R/knitr-engine.R
@@ -21,6 +21,8 @@
 #'
 #' @export
 eng_python <- function(options) {
+  
+  # check for unsupported knitr options
   options <- eng_python_validate_options(options)
 
   # when 'eval = FALSE', we can just return the source code verbatim
@@ -60,6 +62,10 @@ eng_python <- function(options) {
   }
 
   context <- new.env(parent = emptyenv())
+  
+  # environment tracking the labels assigned to newly-created altair charts
+  context$altair_ids <- new.env(parent = emptyenv())
+  
   eng_python_initialize(
     options,
     context = context,
@@ -120,10 +126,10 @@ eng_python <- function(options) {
   had_error <- FALSE
 
   # actual outputs to be returned to knitr
-  outputs <- list()
+  outputs <- stack()
   
   # 'held' outputs, to be appended at the end (for results = "hold")
-  held_outputs <- list()
+  held_outputs <- stack()
 
   # synchronize state R -> Python
   eng_python_synchronize_before()
@@ -142,9 +148,9 @@ eng_python <- function(options) {
     # extract code to be run
     snippet <- extract(code, range)
 
-    # save last value
-    last_value <- py_last_value()
-
+    # clear the last value object (so we can tell if it was updated)
+    py_compile_eval("'__reticulate_placeholder__'")
+    
     # use trailing semicolon to suppress output of return value
     suppress <- grepl(";\\s*$", snippet)
     compile_mode <- if (suppress) "exec" else "single"
@@ -156,15 +162,25 @@ eng_python <- function(options) {
       py_compile_eval(snippet, compile_mode)
 
     # handle matplotlib output
-    captured <- eng_python_matplotlib_handle_output(captured, last_value, i == length(ranges))
+    captured <- eng_python_autoprint(
+      captured   = captured,
+      options    = options,
+      show       = i == length(ranges),
+      context    = context
+    )
 
-    if (length(context$pending_plots) || !identical(captured, "")) {
+    # emit outputs if we have any
+    has_outputs <-
+      !context$pending_plots$empty() ||
+      !identical(captured, "")
+    
+    if (has_outputs) {
 
       # append pending source to outputs (respecting 'echo' option)
       if (!identical(options$echo, FALSE) && !identical(options$results, "hold")) {
         extracted <- extract(code, c(pending_source_index, range[2]))
         output <- structure(list(src = extracted), class = "source")
-        outputs[[length(outputs) + 1]] <- output
+        outputs$push(output)
       }
 
       # append captured outputs (respecting 'include' option)
@@ -174,23 +190,25 @@ eng_python <- function(options) {
           
           # append captured output
           if (!identical(captured, ""))
-            held_outputs[[length(held_outputs) + 1]] <- captured
+            held_outputs$push(captured)
           
           # append captured images / figures
-          for (plot in context$pending_plots)
-            held_outputs[[length(held_outputs) + 1]] <- plot
-          context$pending_plots <- list()
+          plots <- context$pending_plots$data()
+          for (plot in plots)
+            held_outputs$push(plot)
+          context$pending_plots$clear()
           
         } else {
           
           # append captured output
           if (!identical(captured, ""))
-            outputs[[length(outputs) + 1]] <- captured
+            outputs$push(captured)
           
           # append captured images / figures
-          for (plot in context$pending_plots)
-            outputs[[length(outputs) + 1]] <- plot
-          context$pending_plots <- list()
+          plots <- context$pending_plots$data()
+          for (plot in plots)
+            outputs$push(plot)
+          context$pending_plots$clear()
           
         }
 
@@ -217,24 +235,24 @@ eng_python <- function(options) {
   
   if (has_leftovers) {
     leftover <- extract(code, c(pending_source_index, n))
-    outputs[[length(outputs) + 1]] <- structure(
-      list(src = leftover),
-      class = "source"
-    )
+    output <- structure(list(src = leftover), class = "source")
+    outputs$push(output)
   }
   
   # if we were using held outputs, we just inject the source in now
   if (identical(options$results, "hold")) {
     output <- structure(list(src = code), class = "source")
-    outputs[[length(outputs) + 1]] <- output
+    outputs$push(output)
   }
   
   # if we had held outputs, add those in now (merging text output as appropriate)
   text_output <- character()
   
+  held_outputs <- held_outputs$data()
   for (i in seq_along(held_outputs)) {
     
-    if (is.character(held_outputs[[i]])) {
+    output <- held_outputs[[i]]
+    if (!is.object(output) && is.character(output)) {
       
       # merge text output and save for later
       text_output <- c(text_output, held_outputs[[i]])
@@ -243,23 +261,27 @@ eng_python <- function(options) {
       
       # add in pending text output
       if (length(text_output)) {
-        outputs[[length(outputs) + 1]] <- paste(text_output, collapse = "")
+        output <- paste(text_output, collapse = "")
+        outputs$push(output)
         text_output <- character()
       }
       
       # add in this piece of output
-      outputs[[length(outputs) + 1]] <- held_outputs[[i]]
+      outputs$push(held_outputs[[i]])
     }
+    
   }
   
   # if we have any leftover held output, add in now
-  if (length(text_output))
-    outputs[[length(outputs) + 1]] <- paste(text_output, collapse = "")
+  if (length(text_output)) {
+    output <- paste(text_output, collapse = "")
+    outputs$push(output)
+  }
 
   eng_python_synchronize_after()
 
   wrap <- getOption("reticulate.engine.wrap", eng_python_wrap)
-  wrap(outputs, options)
+  wrap(outputs$data(), options)
 
 }
 
@@ -339,7 +361,7 @@ eng_python_initialize_matplotlib <- function(options, context, envir) {
     return()
 
   # initialize pending_plots list
-  context$pending_plots <- list()
+  context$pending_plots <- stack()
 
   plt <- import("matplotlib.pyplot", convert = FALSE)
 
@@ -358,11 +380,11 @@ eng_python_initialize_matplotlib <- function(options, context, envir) {
     graphic <- hook(plt, options)
 
     # update set of pending plots
-    context$pending_plots[[length(context$pending_plots) + 1]] <<- graphic
+    context$pending_plots$push(graphic)
 
     # return None to ensure no printing of output here (just inclusion of
     # plot as a side effect)
-    r_to_py(NULL)
+    py_none()
 
   }
 
@@ -414,26 +436,102 @@ eng_python_is_matplotlib_output <- function(value) {
 
 }
 
-eng_python_matplotlib_handle_output <- function(captured, last_value, show) {
+eng_python_is_altair_chart <- function(value) {
+  
+  # support different API versions, assuming that the class name
+  # otherwise remains compatible
+  classes <- class(value)
+  pattern <- "^altair\\.vegalite\\.v[[:digit:]]+\\.api\\.Chart$"
+  any(grepl(pattern, classes))
+  
+}
 
-  value <- py_last_value()
+eng_python_altair_chart_id <- function(options, ids) {
+  
+  label <- options$label
+  components <- c(label, "altair-viz")
+  if (exists(label, envir = ids)) {
+    id <- get(label, envir = ids)
+    components <- c(components, id + 1)
+    assign(label, id + 1, envir = ids)
+  } else {
+    assign(label, 1L, envir = ids)
+  }
+  
+  paste(components, collapse = "-")
+  
+}
+
+eng_python_autoprint <- function(captured, options, show, context) {
 
   # bail if no new value was produced by interpreter
-  builtins <- import_builtins(convert = TRUE)
-  if (builtins$id(last_value) == builtins$id(value))
+  value <- py_last_value()
+  if (py_is_none(value))
     return(captured)
-
-  # bail if this isn't matplotlib output
-  if (!eng_python_is_matplotlib_output(value))
-    return(captured)
-
-  # show plot if requested
-  if (show) {
-    plt <- import("matplotlib.pyplot", convert = TRUE)
-    plt$show()
+  
+  # ignore placeholder outputs
+  if (inherits(value, "python.builtin.str")) {
+    contents <- py_to_r(value)
+    if (identical(contents, "__reticulate_placeholder__"))
+      return(captured)
   }
 
-  # suppress textual output
-  ""
+  if (eng_python_is_matplotlib_output(value)) {
+    
+    # handle matplotlib output. note that the default hook installed by
+    # reticulate will update the 'pending_plots' item
+    if (show) {
+      plt <- import("matplotlib.pyplot", convert = TRUE)
+      plt$show()
+    }
+    
+    return("")
+    
+  } else if (py_has_attr(value, "_repr_html_")) {
+    
+    # handle plotly output; similarly to above we'll create a plot object
+    # and add that to the context$pending_plots
+    data <- as_r_value(value$`_repr_html_`())
+    context$pending_plots$push(knitr::raw_html(data))
+    return("")
+    
+  } else if (eng_python_is_altair_chart(value)) {
+    
+    browser()
+    
+    # set width if it's not already set
+    width <- value$width
+    if (inherits(width, "altair.utils.schemapi.UndefinedType")) {
+      width <- options$out.width.px %||% 810L
+      value <- value$properties(width = width)
+    }
+    
+    # set height if it's not already set
+    height <- value$height
+    if (inherits(height, "altair.utils.schemapi.UndefinedType")) {
+      height <- options$out.height.px %||% 400
+      value <- value$properties(height = height)
+    }
+    
+    # set a unique id (used for div container for figure)
+    id <- eng_python_altair_chart_id(options, context$altair_ids)
+    
+    # convert to HTML
+    data <- as_r_value(value$to_html(output_div = id))
+    context$pending_plots$push(knitr::raw_html(data))
+    return("")
+    
+  } else if (py_has_attr(value, "to_html")) {
+    
+    data <- as_r_value(value$to_html())
+    context$pending_plots$push(knitr::raw_html(data))
+    return("")
+    
+  } else {
+    
+    # nothing special to do
+    return(captured)
+    
+  }
 
 }